01-ct_training/requirements.txt
google-cloud-aiplatform==1.71.0
google-cloud-pipeline-components==2.17.0
kfp==2.7.0
scikit-learn==1.5.2
mlflow==2.10.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
google-cloud-storage==2.18.2
ml_collections